+++ internal/handler/users.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/OlehHawryliuk/task_manager/internal/model"
 	"github.com/OlehHawryliuk/task_manager/internal/repository"
+	"github.com/OlehHawryliuk/task_manager/internal/service"
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
 )
@@ -18,7 +19,7 @@ func NewUserHandler(repo *repository.UserRepository) *UserHandler {
 	return &UserHandler{repo: repo}
 }
 
-func (h UserHandler) CreateUser(c *gin.Context) {
+func (h *UserHandler) CreateUser(c *gin.Context) {
 	var req model.User
 	err := c.ShouldBindJSON(&req)
 
@@ -29,18 +30,22 @@ func (h UserHandler) CreateUser(c *gin.Context) {
 		return
 	}
 
+	password, err := service.HashPassword(req.Password)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{
+			"error": "Failed to hash password",
+		})
+		return
+	}
+
 	user := &model.User{
 		ID:        uuid.New(),
 		Email:     req.Email,
 		Username:  req.Username,
-		Password:  req.Password,
+		Password:  password,
 		Role:      "user",
 		CreatedAt: time.Now(),
 		UpdatedAt: time.Now(),
-	}
-
-	if req.Role != "" {
-		user.Role = req.Role
 	}
 
 	err = h.repo.CreateUser(user)
@@ -55,17 +60,45 @@ func (h UserHandler) CreateUser(c *gin.Context) {
 }
 
 func (r *UserHandler) GetUserByID(c *gin.Context) {
+	currentUserID := c.GetString("userID")
+	ParsedCurrentUserID, err := uuid.Parse(currentUserID)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid userID",
+		})
+		return
+	}
+
 	id := c.Param("id")
 
 	userID, err := uuid.Parse(id)
 	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid user ID"})
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid user ID",
+		})
+		return
+	}
+
+	currentUser, err := r.repo.GetUserByID(ParsedCurrentUserID)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "User not found",
+		})
+		return
+	}
+
+	if ParsedCurrentUserID != userID && currentUser.Role != "admin" {
+		c.JSON(http.StatusForbidden, gin.H{
+			"error": "Forbidden: you can only view ypur own profile",
+		})
 		return
 	}
 
 	user, err := r.repo.GetUserByID(userID)
 	if err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Failed to fetch user"})
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "User not found",
+		})
 		return
 	}
 
@@ -92,18 +125,36 @@ func (r *UserHandler) GetUserByEmail(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{
 			"error": err.Error(),
 		})
+		return
 	}
 
 	c.JSON(http.StatusOK, user)
 }
 
 func (r *UserHandler) UpdateUser(c *gin.Context) {
+	currentID := c.GetString("userID")
+	if currentID == "" {
+		c.JSON(http.StatusUnauthorized, gin.H{
+			"error": "Missing user ID",
+		})
+		return
+	}
+
+	parsedCurrentUserID, err := uuid.Parse(currentID)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid user ID",
+		})
+		return
+	}
+
 	id := c.Param("id")
 	userID, err := uuid.Parse(id)
 	if err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{
 			"error": "Faild to parse user ID",
 		})
+		return
 	}
 
 	var req model.User
@@ -112,11 +163,29 @@ func (r *UserHandler) UpdateUser(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{
 			"error": err.Error(),
 		})
+		return
+	}
+
+	currentUser, err := r.repo.GetUserByID(parsedCurrentUserID)
+	if err != nil {
+		c.JSON(http.StatusNotFound, gin.H{
+			"error": "User not found",
+		})
+		return
+	}
+
+	if parsedCurrentUserID != userID && currentUser.Role != "admin" {
+		c.JSON(http.StatusForbidden, gin.H{
+			"error": "Forbidden: You can only update your own profile",
+		})
+		return
 	}
 
 	user, err := r.repo.GetUserByID(userID)
 	if err != nil {
-		c.JSON(http.StatusNotFound, gin.H{"error": "User not found"})
+		c.JSON(http.StatusNotFound, gin.H{
+			"error": "User not found",
+		})
 		return
 	}
 
@@ -127,14 +196,23 @@ func (r *UserHandler) UpdateUser(c *gin.Context) {
 		user.Username = req.Username
 	}
 	if req.Password != "" {
-		user.Password = req.Password
+		hashedPassword, err := service.HashPassword(req.Password)
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, gin.H{
+				"error": "Failed to hash password",
+			})
+			return
+		}
+		user.Password = hashedPassword
 	}
-	if req.Role != "" {
+	if req.Role != "" && user.Role == "admin" {
 		user.Role = req.Role
 	}
 
 	if err := r.repo.UpdateUser(user); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update user"})
+		c.JSON(http.StatusInternalServerError, gin.H{
+			"error": "Failed to update user",
+		})
 		return
 	}
 
@@ -142,6 +220,15 @@ func (r *UserHandler) UpdateUser(c *gin.Context) {
 }
 
 func (r *UserHandler) DeleteUser(c *gin.Context) {
+	currentID := c.GetString("userID")
+	ParsedCurrentID, err := uuid.Parse(currentID)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": "Invalid user ID",
+		})
+		return
+	}
+
 	id := c.Param("id")
 	userID, err := uuid.Parse(id)
 
@@ -149,6 +236,22 @@ func (r *UserHandler) DeleteUser(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{
 			"error": err.Error(),
 		})
+		return
+	}
+
+	user, err := r.repo.GetUserByID(ParsedCurrentID)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{
+			"error": "User not found",
+		})
+		return
+	}
+
+	if userID != ParsedCurrentID && user.Role != "admin" {
+		c.JSON(http.StatusForbidden, gin.H{
+			"error": "Forbidden: You can only delete your own profile",
+		})
+		return
 	}
 
 	err = r.repo.DeleteUser(userID)
@@ -156,6 +259,7 @@ func (r *UserHandler) DeleteUser(c *gin.Context) {
 		c.JSON(http.StatusBadRequest, gin.H{
 			"error": "Failed to delete user",
 		})
+		return
 	}
 
 	c.JSON(http.StatusOK, gin.H{
